model: move Job.WriteConfigs to RoleJob
This means we don't attempt to put the links from other jobs into the
config file.

Also, drop the useless .job.templates key from config_spec.json.

Addresses:
- https://github.com/SUSE/fissile/pull/288#discussion_r146407701
- https://github.com/SUSE/fissile/pull/288#discussion_r146407497
- https://github.com/SUSE/fissile/pull/288#discussion_r146406839
- https://github.com/SUSE/fissile/pull/288#discussion_r146407249
- https://github.com/SUSE/fissile/pull/288#discussion_r146423803
- https://github.com/SUSE/fissile/pull/288#discussion_r146423128
